perf(migration): add indexes for findChildrenTree
Add indexes for the most frequent query in the product environment.

We're trying to aviod seq scans and hit indexes every time if possible.

WARNING: This migration may lock the target tables. We should create
the index concurrently but there is an implementation  detail in the
node-pg-migrate that forces a transaction for each pgm.dropIndex call
and postgresql does not allow it. At the current time, the database is
not huge and we do not expect a heavy system load at this time.

Links:

- https://github.com/filipedeschamps/tabnews.com.br/issues/879
- https://explain.depesz.com/s/OIKV
- https://github.com/salsita/node-pg-migrate/issues/709