refactor(request builder): use consistent interfaces and response body parsing
The decision on how to treat the response in the "act" step of a test is best made by the test
itself, and not by the `RequestBuilder`, which is just an abstraction to make writing and reading
tests easier. Also, the interface should be consistent accross `get`, `patch` and `post` to make
it easier to use and understand the `RequestBuilder` class.