chore(middleware): disable active middleware
This commit concludes the process of removing the active middleware from the
application. Its responsibilities were migrated or delegated in previous commits.

The remaining logic for "under maintenance" mode has been disabled, and its tests
have been skipped. The file was renamed to `middleware.private.js` to serve as a
placeholder, allowing for easy reactivation in the future if a new need for a
middleware arises.

Co-authored-by: Rafael Tavares <26308880+Rafatcb@users.noreply.github.com>